..
git-svn-id: http://svn.l2jfree.com/l2jfree@6256 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/l2jfree-core/src/main/java/com/l2jfree/gameserver/network/clientpackets/L2GameClientPacket.java
@@ -143,7 +143,7 @@ public abstract class L2GameClientPacket extends ReceivablePacket<L2GameClient>
 	protected final int readCompQ()
 	{
 		if (Config.PACKET_FINAL)
-			return L2Math.limit(Integer.MIN_VALUE, readQ(), Integer.MAX_VALUE);
+			return (int) L2Math.limit(Integer.MIN_VALUE, readQ(), Integer.MAX_VALUE);
 		else
 			return readD();
 	}
